+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Toon een indicator of er een raportje is gemaakt of een bericht is verstuurd in een report waar jij deel van maakt
 - Restore health and armor on join
 - Copy VIN to clipboard when reading it
+- Je kan nu minimap zien in noclip
 
 ### Changed